==> vignettes/5.SNV_correction.Rmd <==
---
title: "5.SNV correction"
output:
  html_document:
    df_print: paged
vignette:  >
  %\VignetteIndexEntry{5.SNV correction} 
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
lang: "en-US"
---

```{r, include = FALSE}
knitr::opts_chunk$set(
  echo=TRUE,
  progress =FALSE,
  comment = "#>"
)
options(rmarkdown.html_vignette.check_title = FALSE)
```

We have conducted benchmark of several SNV denoising tools designed for single-cell data, and found that CBM show superior performance (please refer to our paper for more details). Users should install and run [CBM](https://github.com/zhyu-lab/cbm) before running MAAS. \

In this tutorial, CBM has been installed in `~/App/cbm-main`.

```{r, message=FALSE, warning=FALSE, eval=FALSE}
library(dplyr)
library(data.table)
library(stringr)

input_file <- "snv.CBMinput.txt"
```

The raw SNV profile as input for CBM is a cell-by-mutation matrix (cells in rows and mutations in columns). The cell and mutation names will be recovered after executing CBM.

``` {r, message=FALSE, warning=FALSE, eval=FALSE}
input <- read.table("http://bioinfo.szbl.ac.cn/share/MAAS_data/snv.CBMinput.txt", sep = '\t', header = T)
input[1:10,1:10]
```

``` {r, echo=FALSE, message=FALSE, warning=FALSE}
x <- read.table("http://bioinfo.szbl.ac.cn/share/MAAS_data/snv.CBMinput.txt", sep = '\t', header = T)
x[1:10,1:10]
```

``` {r, message=FALSE, warning=FALSE, eval=FALSE}
# sample.tmp <- gsub("\\.snv\\.CBMinput\\.txt", "", file.tmp)
output_file <- "./"
cmd <- paste("~/App/cbm-main/bin/cbm -i", input_file, "-o", output_file, "-K 10") # Set up to 10 clones
system(cmd)
```

``` {r, message=FALSE, warning=FALSE, eval=FALSE}
pred.cluster.genotype <- as.data.frame(fread(".clones.txt"))
rownames(pred.cluster.genotype) <- 0:(nrow(pred.cluster.genotype)-1)
pred.cluster.genotype[,1:10]
```

The genotype of each cluster. Clusters in rows and mutations in columns
``` {r, echo=FALSE, message=FALSE, warning=FALSE}
pred.cluster.genotype <- read.table("http://bioinfo.szbl.ac.cn/share/MAAS_data/.clones.txt", sep = '\t', header = F, check.names = F)
rownames(pred.cluster.genotype) <- 0:(nrow(pred.cluster.genotype)-1)
pred.cluster.genotype[,1:10]
```

```{r, message=FALSE, warning=FALSE, eval=FALSE}
pred.cluster.label <- fread(".assignment.txt")
pred.cluster.label <- as.data.frame(t(pred.cluster.label))
colnames(pred.cluster.label) <- "Cluster" # One-dimension vector: Cells in rows with assigned clusters
pred.cluster.label[1:10,,drop = F]
```

``` {r, echo=FALSE, message=FALSE, warning=FALSE}
pred.cluster.label <- data.table::fread("http://bioinfo.szbl.ac.cn/share/MAAS_data/.assignment.txt")
pred.cluster.label <- as.data.frame(t(pred.cluster.label))
colnames(pred.cluster.label) <- "Cluster" # One-dimension vector: Cells in rows with assigned clusters
pred.cluster.label[1:10,,drop = F]
```

The cells from the same clone are assumed to be same. Hence, we assign cluster-level mutations to individual cells
``` {r, message=FALSE, warning=FALSE, eval=FALSE}
pred <- do.call(rbind, lapply(pred.cluster.label$Cluster, function(cluster) {
  pred.cluster.genotype[as.character(cluster), ]
}))
pred[1:10,1:10]
```

```{r, echo=FALSE, message=FALSE, warning=FALSE}
x <- read.table("http://bioinfo.szbl.ac.cn/share/MAAS_data/denoised.mat.txt", sep = '\t', header = T, row.names = 1, check.names = F)
x[1:10,1:10]
```

``` {r, message=FALSE, warning=FALSE, eval=FALSE}
raw.snv <- readRDS("snv.mat.rds")
rownames(pred) <- colnames(raw.snv)
colnames(pred) <- rownames(raw.snv)
saveRDS(pred, "denoised.mat.rds")
```
